fix: Rename NuruCommand to NuruMediatorCommand and implement proper mediator pattern
- Renamed NuruCommand.cs to NuruMediatorCommand.cs for clarity
- Implemented proper mediator pattern with nested handler using IRequest/IRequestHandler
- Fixed AddDependencyInjection to scan assembly for handlers
- Updated benchmark to use new class name
